fix 3

+++ include/task2.h
@@ -1,5 +1,4 @@
-
-template<typename T, size_t size>
+template<typename T, int size>
 T* createArr(T(*gen)()) {
     T* arr = new T[size];
     for (int i = 0; i < size; i++)
